Accommodate rows that lack specified field
SharePoint list items that don’t have a value for a field are missing
that field altogether in the GetListItems results. Use get on the row’s
inner dictionary to avoid throwing an error in such a case (defaults to
None). Then, filter out the None keys from the mapping table.